RT#76038 Enforce Test::EOL loading
RT#76038: https://rt.cpan.org/Ticket/Display.html?id=76038#txn-1057838

Before this fix, the test succeeds (being skipped) if Test::EOL
is missing. This means that the test succeeds even if there are
some EOL problems somewhere.

This patch replace conditional loading with Test::Requires by
unconditional loading.
This is fine because the test is run only when AUTHOR_TESTING is set,
and because Test::EOL is a requirement of DZP::Test::EOL.